FilmPage & fecthDrupalData: optimize logics (lint)
Refactors the film page to improve data display when specific fields are missing, providing a better user experience.

Moves the `fetchDrupalFilms` function inside the `useEffect` hook in `useFetchAllFilms` to prevent unnecessary re-renders.

Adds `initialData` to the dependency array of the `useEffect` hook in `useFetchAllFilms` to ensure the hook is triggered when initial data changes.