searchPage WIP : Adds search functionality
Implements a search page with a dedicated film grid.
This change introduces a `SearchPage` component that displays film results based on a search term (slug).

The `FilmsGrille` component is updated to accept a `searchTerm` prop, allowing it to filter films based on the search query. The `useFetchAllFilms` hook is also updated to take this term into account and display relevant results

A Thematic filter is now hidden when performing a search