filmsGrille: improves order randomization & adds isRealted option
Moves the random style creation logic to a utility function for better code organization and reusability.

Improves the randomization algorithm to ensure a more uniform distribution of film card styles.

Adds a condition to exclude the "ThematiqueFilter" component and the films counter from the "related films" section.

Ensures that `useEffect` hook is run when `isRelated` prop changes.