Page non trouvée personnalisée : globale et pour les page de film unique
Implements a 404 page for the entire app and  it when a film is not found.

This change ensures that users are presented with a friendly "Page not found" message when they attempt to access a film that doesn't exist, improving the overall user experience.

The logic now checks if a film is found before rendering the film page. If not, it triggers the `notFound()` function, which renders the custom 404 page.

Updates the data fetching logic to handle cases where a film is not found.