Adds support for field_resume_de_l_institution_de
Ajustement # 11
Adds the "institution description" field to film pages and search results.

This change allows users to search and view films based on the description provided by the institution. It enhances the search functionality by including this new field in the search index and displays it on the film's detail page.